+++ yash-env/src/builtin.rs
@@ -66,11 +66,10 @@ pub enum Type {
     /// Search and Execution](https://pubs.opengroup.org/onlinepubs/9799919799/utilities/V3_chap02.html#tag_19_09_01_04)
     /// in POSIX XCU section 2.9.1.4.
     /// They are very similar to mandatory built-ins, but their behavior is not
-    /// specified by POSIX, so they are not portable. They cannot be used when
-    /// the (TODO TBD) option is set. <!-- An option that disables non-portable
-    /// behavior would make elective built-ins unusable even if found. An option
-    /// that disables non-conforming behavior would not affect elective
-    /// built-ins. -->
+    /// specified by POSIX, so they are not portable. When the
+    /// [`Portable`](crate::option::Portable) option is on, attempting to
+    /// execute one is rejected with an error, even though it is still found in
+    /// command search.
     ///
     /// We call them "elective" because it is up to the shell whether to
     /// implement them.
@@ -88,8 +87,9 @@ pub enum Type {
     /// - When the [`PosixlyCorrect`](crate::option::PosixlyCorrect) option is
     ///   on, they are ignored: they are regarded as non-existing utilities so
     ///   that the command search falls through to external utilities.
-    /// - When the (TODO TBD) option is on, they cannot be used even if found
-    ///   in command search.
+    /// - When the [`Portable`](crate::option::Portable) option is on,
+    ///   attempting to execute one is rejected with an error, even though it
+    ///   is still found in command search.
     Extension,
 
     /// Built-in that works like a standalone utility